CORE:1256: Stellar fixes (from real world test)
Fix compiler warnings as well

+++ WalletKitCore/src/walletkit/handlers/xlm/WKSupportXLM.c
@@ -19,7 +19,7 @@ private_extern BRCryptoAmount
 cryptoAmountCreateAsXLM (BRCryptoUnit unit,
                          BRCryptoBoolean isNegative,
                          BRStellarAmount value) {
-    return cryptoAmountCreate (unit, isNegative, uint256Create (value));
+    return cryptoAmountCreate (unit, isNegative, uint256Create ((uint64_t)value));
 }
 
 private_extern BRCryptoHash
